requirements.txt

flask==2.0.2
requests
gunicorn==20.1.0
openpyxl
python-decouple